requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pyjwt
sqlalchemy
pyodbc
